Fixed problem with xgboost going to 1.6.0
I don't know when this happened, but probably within the last month and it broke things. It does highlight that, in theory, we might be able to install an older version of xgboost to match what was used, but the updated warning from 1.5.2 (vs 1.5.0) makes me feel more comfortable that it's still working, it's just not guaranteed to continue working).

I also remvoed the mview install directory, which I think was safe... I'll find out when I come in to a hopefully non-broken build tomorrow morning.